修改超参数使得老师模型的asr符合预期

+++ src/trainer/kdbackdoor.py
@@ -14,10 +14,10 @@ from src.config import base_config
 
 if __name__ == "__main__":
     datamodule_name = "cifar10"
-    epochs = 30
-    lr = 0.1
-    poison_rate = 0.2
-    teacher_network = "mobilenetv2"
+    epochs = 100
+    lr = 0.001
+    poison_rate = 0.01
+    teacher_network = "resnet18"
 
     # pretrain teacher model
     pretrain_model = NormalModel(
@@ -37,17 +37,17 @@ if __name__ == "__main__":
         gpus=1,
         auto_select_gpus=True
     )
-    pretrain_trainer.fit(
-        model=pretrain_model,
-        datamodule=pretrain_model.datamodule
-    )
-    # HACK
-    # ugly hack
     pretrain_model_path = pretrain_checkpoint_dir / \
-        f"{teacher_network}-{datamodule_name}-epochs={epochs}-lr={lr}-pretrain.pt"
-    if not os.path.exists(pretrain_model_path.parent):
-        os.makedirs(pretrain_model_path.parent)
-    torch.save(pretrain_model._network.state_dict(), str(pretrain_model_path))
+        f"{teacher_network}-{datamodule_name}-adam-epochs={epochs}-lr={lr}-pretrain.pt"
+    
+    if not os.path.exists(pretrain_model_path):
+        pretrain_trainer.fit(
+            model=pretrain_model,
+            datamodule=pretrain_model.datamodule
+        )
+        if not os.path.exists(pretrain_model_path.parent):
+            os.makedirs(pretrain_model_path.parent)
+        torch.save(pretrain_model._network.state_dict(), str(pretrain_model_path))
 
     # kdbackdoor
     model = KDBackdoorModel(
